refactor: align MTD components with backend module layout
使 MTD 组件符合后端模块布局

class DiarizationSegment(TypedDict):
    start_s: float
    end_s: float
    speaker_id: str
    text: str

在 speaker_diarization 接口中明确公共分段类型，并由两个模型客户端返回统一结构。

class _SpeakerExemplarPool:
    ...

def _build_audio_layout(...):
    ...

将仅属于会话调度的 exemplar、音频布局和时间线处理收回 MtdDiarizationManager，删除 serving/mtd 中间包，同时重命名并更新对应测试。
