+++ src/xtalk/models/speaker_diarization/__init__.py
@@ -1,11 +1,12 @@
 """Speaker-diarization model implementations."""
 
-from .interfaces import DiarizationResult, SpeakerDiarization
+from .interfaces import DiarizationResult, DiarizationSegment, SpeakerDiarization
 from .official_mtd_client import OfficialMtdClient
 from .sglang_omni_mtd_client import SglangOmniMtdClient
 
 __all__ = [
     "DiarizationResult",
+    "DiarizationSegment",
     "OfficialMtdClient",
     "SglangOmniMtdClient",
     "SpeakerDiarization",
