--- src/xtalk/serving/mtd/__init__.py
@@ -1,15 +1,0 @@
-"""MTD serving helpers."""
-
-from .audio_layout import MtdAudioLayout, build_audio_layout
-from .exemplar_pool import SpeakerExemplar, SpeakerExemplarPool
-from .transcript import DiarizationSegment, parse_segments, render_segments
-
-__all__ = [
-    "DiarizationSegment",
-    "MtdAudioLayout",
-    "SpeakerExemplar",
-    "SpeakerExemplarPool",
-    "build_audio_layout",
-    "parse_segments",
-    "render_segments",
-]
